+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased ([changes](https://github.com/infertux/bashcov/compare/v2.0.0...master))
 
+  * [BUGFIX]  Ensure that `SimpleCov.result?` returns `true` at process exit so
+              that (e.g.) SimpleCov causes `bashcov` to exit with nonzero
+              status if the coverage percentage is less than
+              `SimpleCov.minimum_coverage`.
   * [FEATURE] Add `$BASHCOV_ROOT` environment variable that corresponds to the
               `--root` command-line  and `Bashcov.root_directory` setting.
   * [FEATURE] Add utility methods for loading SimpleCov with and without